refactor(cli): rename formatter.rs -> tui_formatter.rs for naming consistency
The 'formatter.rs' module exposes a 'Formatter' struct that builds
ratatui Lines + Spans for the TUI. Despite living next to tui.rs,
tui_export.rs, and tui_settings.rs (all 'tui_*'-prefixed), it was
named just 'formatter.rs', creating apparent duplication with the
sibling 'cli_formatter.rs' (which is the OTHER formatter — for plain
ANSI CLI output via the CliFormatter struct).

This is a pure rename. No type signature changes, no behavior
changes, just:
- mv formatter.rs -> tui_formatter.rs
- mod formatter; -> mod tui_formatter; in main.rs
- use crate::formatter::Formatter -> use crate::tui_formatter::Formatter
  in main.rs and tui.rs

After this: tui*-prefixed files form a coherent group:
  tui.rs, tui_export.rs, tui_formatter.rs, tui_settings.rs

And the CLI/TUI formatter split is unambiguous from the file name.
