+++ apps/netscli-cli/src/tui_formatter.rs
